Refactored DepthwiseConv2D and other corrections
